+++ src/main/java/org/example/models/Usuario.java
@@ -16,7 +16,7 @@ public class Usuario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id_usuario;
+    private int id_usuario;
 
     private String nome;
     private String email;
